FP% and FPW with decimals as mouse-over
Mouse-over might be useful for more people, yes. 

With some help from a friend I got that to work, currently the mouse-over has 4 decimals for FP% and 6 decimals for FPW. Not sure if this should be considered an easter egg or if a .toFixed(2) should be added to make it match the presentation in the top lists? The actual FP% and FPW top lists show two decimals to users. 

Also added Math.round() to the shown part of the module, this should make the rounding more correct instead of the current parseInt that just drops the decimals (ie with Math.round() it will match the mouse-over from the toplists at PGC better and be more correct). My understanding is that this is a faster/better solution than .toFixed(0) for whole numbers but feedback appreciated if I've misunderstood something. Goal is more consistency with the data and less risk for confusion for the user.